fix(frontend): Passing icon & text from SSR to client for nav admin
